32 (#34)
* Removing bad config

* Moving DB outside of app as an experiment. Adding optional chaining to cache getters to fix issue in build.
