+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import Text from "@/components/text/text";
 import { notFound } from "next/navigation";
-import { getDocument } from "./db/db";
+import { getDocument } from "../db/db";
 import type { Config, Text as TextType } from "@/lib/config-provider";
 
 export default async function Home() {
