+++ src/instrumentation.ts
@@ -2,7 +2,7 @@ import type NodeCache from 'node-cache';
 
 export async function register() {
   if (process.env.NEXT_RUNTIME === 'nodejs') {
-    const NodeCache = (await import('node-cache')).default;
+    const { default: NodeCache } = await import('node-cache');
     const config: NodeCache.Options = {
       stdTTL: process.env.NODE_ENV === 'production' ? 0 : 60,
     };
